Apis/risk_score/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.1

scikit-learn==1.7.2
lightgbm==4.2.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2

python-dotenv==1.0.1
pydantic==2.6.0
pyyaml==6.0.1

pyarrow==15.0.0

matplotlib==3.8.2
fpdf==1.7.2

pocketbase==0.11.0
